+++ common/api/summary/core-frontend.exports.csv
@@ -130,7 +130,6 @@ internal;function;createDefaultViewFlagOverrides
 internal;function;createGraphicFromDescription
 beta;interface;CreateGraphicFromDescriptionArgs
 beta;interface;CreateGraphicFromTemplateArgs
-internal;function;createGraphicTemplate
 internal;function;createGraphicTemplateFromDescription
 internal;function;createMapLayerTreeReference
 internal;function;createMaskTreeReference
@@ -313,9 +312,6 @@ public;interface;GraphicShape
 public;interface;GraphicShape2d
 public;interface;GraphicSolidPrimitive
 beta;interface;GraphicTemplate
-internal;interface;GraphicTemplateBatch
-internal;interface;GraphicTemplateBranch
-internal;interface;GraphicTemplateNode
 public;enum;GraphicType
 internal;interface;GroundPlaneDecorations
 public;interface;GroupMark
