+++ core/frontend/src/internal/render/PrimitiveBuilder.ts
@@ -17,7 +17,8 @@ import { MeshList } from "../../common/internal/render/MeshPrimitives";
 import { GraphicBranch } from "../../render/GraphicBranch";
 import { assert } from "@itwin/core-bentley";
 import { _accumulator, _implementationProhibited } from "../../common/internal/Symbols";
-import { createGraphicTemplate, GraphicTemplate, GraphicTemplateBatch } from "../../render/GraphicTemplate";
+import { createGraphicTemplate, GraphicTemplateBatch } from "../../internal/render/GraphicTemplateImpl";
+import { GraphicTemplate } from "../../render/GraphicTemplate";
 import { RenderGeometry } from "./RenderGeometry";
 
 export class PrimitiveBuilder extends GraphicBuilder {
